Revert "expire and restore metric paths"
