+++ dlcli/cli/restore.py
@@ -9,7 +9,7 @@ logger = logging.getLogger(__name__)
 
 @cli.group('restore')
 def restore():
-    """restores backups and metric paths"""
+    """restores backups"""
 
 
 @click.command(short_help="Restore an Account")
@@ -34,16 +34,5 @@ def org(org):
         sys.exit(1)
 
 
-@click.command(short_help="Restore metric paths")
-def metrics():
-    try:
-        resp = series.update_tag_metric_paths(tag="all", **context.settings)
-        click.echo(click.style('Recovered: %s paths', fg='green') % (resp))
-    except Exception, e:
-        print 'Cleanup metrics failed. %s' % e
-        sys.exit(1)
-
-
 restore.add_command(account)
 restore.add_command(org)
-restore.add_command(metrics)
